feat: added authservices -> forgot-password, verify-otp, reset-password

+++ src/services/user.service.ts
@@ -2,7 +2,7 @@ import { Prisma, UserRole } from "@prisma/client";
 import { UserResponse } from "../models/user.model";
 import { UserRepository } from "../repositories/user.repository";
 import { Token } from "../types/token.type";
-import { InternalServerError, NotfoundError } from "../utils/errors";
+import { BadrequestError, InternalServerError, NotfoundError } from "../utils/errors";
 import responses from "../utils/responses";
 
 export class UserService {
@@ -25,10 +25,13 @@ export class UserService {
 		return responses.toUserResponses(result);
 	}
 
-	static async getById(id: string): Promise<UserResponse> {
+	static async getById(id: string, user: Token): Promise<UserResponse> {
 		const result = await UserRepository.findById(id);
 
 		if (!result) throw new NotfoundError(`Pengguna tidak ditemukan`);
+
+		if (user.role !== "ADMIN" && user.role !== "STAFF")
+				if (result.role === "ADMIN" || result.role === "STAFF") throw new BadrequestError("Pengguna tidak ditemukan")
 
 		return responses.toUserResponse(result);
 	}
@@ -37,6 +40,8 @@ export class UserService {
 		const checkUser = await UserRepository.findById(id);
 
 		if (!checkUser) throw new NotfoundError("Pengguna tidak ditemukan");
+
+		if(checkUser.role === "ADMIN") throw new NotfoundError("Pengguna tidak dapat di hapus")
 
 		const result = await UserRepository.deleteById(checkUser.id);
